+++ lib/redux.rb
@@ -5,7 +5,13 @@ module Redux
   def self.combine_reducers(reducers)
     ->(state = {}, action){
       reducers.reduce({}){ |next_state, (key, reducer)|
-        next_state[key] = reducer.call(state[key], action)
+        if state.key?(key)
+          next_state[key] = reducer.call(state[key], action)
+        else
+          next_state[key] = reducer.call(action)
+        end
+
+        next_state
       }
     }
   end