+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## CHANGELOG
 
+### Next
+
+* Redux.combine_reducers
+
 ### 0.1.0
 
-* Inital release
+* Initial release